Variable sizeText

+++ examples/PIO_TestPatterns/src/main.h
@@ -10,4 +10,4 @@ uint16_t XY16( uint16_t x, uint16_t y);
 void mxfill(CRGB *leds);
 uint16_t colorWheel(uint8_t pos);
 void drawText(const String& printedText);
-void spacingAngelDisplay(double distance, double speed);
+void spacingAngelDisplay(int distance, double speed);
